RAINGAUGE/requirements.txt
requests==2.24.0
pyserial==3.4

